backend/requirements.txt
Django==5.1.7
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
django-filter==24.3
psycopg2-binary==2.9.10
python-dotenv==1.0.1
Pillow==10.4.0
pandas==2.2.3
openpyxl==3.1.5
reportlab==4.2.5
requests==2.32.3
gunicorn==21.2.0
